AI_News_Telegram_Bot/requirements.txt
requests>=2.31.0
feedparser>=6.0.0
openai>=1.0.0
python-dotenv>=1.0.0
